fix(Makefile): always update embedded binary data
I have find it very surprising that this is not done automatically
when creating a new release. I fixed the SQL statements that were
preventing the registry from going up (and that was confirmed by a
Postgres syntax checking tool like markdrago/pgsanity), yet when
I run the registry on the server I was seeing the error. I was like
"WTF?!", then reading the Makefile I finally figured out that the
problem is that changing binary data did not lead to the _embedded_
binary data to be updated. Since expressing all the binary data
as dependencies is error prone, I decided to _unconditionally_ and
_always_ run the binary data embedding code. At least, with this
strategy it cannot happen what led me to loose time and to become
very frustrated because I did not understood what was going on.