Get the locally generate key and certificate to be written to the card
It seems like you need to have a certificate associated to the private
key in order for the private key to be written.
For the time being I added a certificate with a expiry date of 10 years.

It took me a fair amount of time to figure this out and the yubikey docs
don't really help that much.

Tooling I have found useful to debug this problem are:
- pkcs11-tool --module /usr/local/lib/libykcs11.dylib -O --so-pin XXXX
- p11tool
- yubico-piv-tool